move function and error definition to api module
